services/api-gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
